msm: cpr-regulator: Enable CPR after both voltage and enable state are set
Change the cpr-regulator driver so that RBCPR hardware is only
enabled after consumers have called both regulator_set_voltage()
and regulator_enable().  Currently, RBCPR is enabled as soon as
a voltage (corner) is specified.  Unfortunately, the physical
regulator voltage does not change until regulator_enable() is
also called.  This leads to the RBCPR hardware recommending that
the voltage be stepped up even though it shouldn't be.

Change-Id: I9c545245e41cdd813cc3425ef789c8bff13898fa
Signed-off-by: David Collins <collinsd@codeaurora.org>