qseecom: Fix return path of qseecom load app if app exists
Fixes the issue during the return path of qseecom load_app. If
the application is already up and running then second time load
app should return with proper error. During this process SFAB
disable clock vote is missed, which might create problem. So
this patch adds the appropriate disable call for the SFAB clock
which fixes the issue.

Change-Id: I744a923f7e08fea85ef5bada5538f1a28c52e4e7
CRs-fixed: 498153
Signed-off-by: AnilKumar Chimata <anilc@codeaurora.org>