msm: mdss: keep reference of rotator output buffer
Change logic to free buffer for previous rotation call. This will make
sure to keep a ref count on the buffer secure lock before it's picked up
by MDP for displaying to panel. It will also allow for asynchronous calls
to rotator to be supported by not freeing the source and destination
buffers within same call.

Change-Id: I389d6a4523bdc79a5dfe76e3603e43325d63b82c
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>