usb: usb_bam: Don't wait for consumer request on connect pipes
On certain compositions, ECM for instance, consumer is not requested
on pipe connection and thus the part of the handshake that waits
for consumer request causes a timeout and a delay of 10 seconds
before network connection can be established.
In a revisited design for IPA-USB resource handshake,
USB shall not wait for USB_CONS resource, thus avoiding the
delay on ECM compositions.

CRs-fixed: 499480
Change-Id: I2a7e44289c13b5e2f8f69c9c26f5af4b6eba510c
Signed-off-by: Lena Salman <esalman@codeaurora.org>