ARM: dts: msm: Remove USB3 ID IRQ from apq8074-dragonboard
commit 4db88d98 (ARM: dts: msm: Move USB3 interrupt overrides
back to top-level msm8974) refactored the &usb3 interrupt map.
The apq8074-dragonboard DTSI is also affected downstream as it
includes msm8974.dtsi. However, on this platform the &usb3
device does not use the pmic_id_irq but rather the &usb_otg
device does. Thus we need to add an override to remove the
pmic_id_irq from &usb3 on this platform.

Change-Id: I22073913be3ab56e3894d1351804973bcbd2c6cb
Signed-off-by: Jack Pham <jackp@codeaurora.org>