msm: mdss: enable continuous splash screen for 8x10
Continuous splash screen is to allow lk to kernel smooth
transition for the display splash screen.  During lk to
kernel transition, the splash image will be copied and
display engine status will be preserved so that user will
see continuous splash screen.   This is to enable the
feature on 8x10 MTP and CDP platforms.

Change-Id: Ie8f71c1d0400cb72dff3a076f57c1d7e19a844f0
Signed-off-by: Terence Hampson <thampson@codeaurora.org>
Signed-off-by: Xiaoming Zhou <zhoux@codeaurora.org>