msm: vidc: add set ctrl for Dynamic Buffer mode
The video driver will support Dynamic Buffer mode
for capture port. With Static Buffer mode, all the
output buffers are pre-registered with firmware
before stream-on. In dynamic buffer mode, there will
be no pre-registeration of output buffers with
firmware. This mode is primarily for streaming
use case, to save on video start-up latency.

Change-Id: Iafb9541bd185a474b169dfe17a5a278b930bdad6
Signed-off-by: Arun Menon <avmenon@codeaurora.org>