ARM: dts: msm: Configure regulator LDO 4 to be always enabled for msm8610
The correct sequence for audio codec power-on is for the 1.2 V codec
analog supply, PM8110 LDO 4, to be enabled before the codec is brought
out of reset. Once the codec is reset, LDO 4 should never be disabled.
If LDO 4 is disabled, then the audio codec will be in an undefined
state which causes both noise in the speaker and the codec I2C
slave to not respond to the I2C master. Configure the LDO 4 regulator
to be always enabled in order to eliminate the issue. Keeping LDO 4
on will not impact the MSM8610 power target.

Change-Id: I9c59a1965178c04333665b48f847002eb587c8a3
CRs-Fixed: 500003
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>