msm: clock-local2: Increase settling delays for CLKFLAG_RETAIN_*
8974v2.2, Pro, and future devices add small hardware delays between
powering on individual memory banks in an effort to minimize voltage
droops. Consequently, the memories will take slightly longer to enable
(200ns to 400ns, depending on the chip) than on earlier hardware.

Replace the current delay (achieved with dummy register accesses)
with an explicit 1us delay to support the newer chips.

Change-Id: Ib69812a50edf3f700e802057034f85a1bbcef43f
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>