msm: gdsc: Remove csi_clk usage in the 8x10's VFE GDSC sequence
The CSI_VFE_CLK does not enter the VFE's power-collapsible domain,
and its reset therefore does not need to be controlled. Remove this
unnecessary usage so that the resets asserted manually (when the
'qcom,skip-logic-collapse' GDSC property is set) are the same as
those asserted by the GDSC state machine when the property is not
set.

Change-Id: I109e75ae85ee137f57d4c33ab191d0409f271ef5
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>